Committed v0.17
Allowed the number of a fixed value to be set to zero (i.e., that value will not appear in the generated sample).
Allowed decimal places to be set to zero.
Added link to GitHub.
Fixed a bug that could cause the display to be distorted if the X-axis started at a value above 0.
Removed bounds on scale maximum/minimum input values. Caveat usor.
Fixed a bug that could cause no solution to be found when fixed values were used and there was a small number of possible solutions.